First attempt at range checks on squaring
This is based on the realisation that over half of the entries in the
squares table are overflows, where the table just returns the largest
representable number, so with a couple of range checks those could be
omitted from the table.  Also the realisation that any iteration that
hits one of those will (I think) be the last one, so having detected an
overflow, we can just bail out.

Unfortunately, the range checks are needed in three places, and that's
enough to make the kernel too big for int zero-page slot, so only two of
them are turned on for now.