Almost-working upgrade to 15-bit precision
This uses range-checking to arrange that we only store squares of
numbers between -4 and +4.  Outside that range if we need the square of
a number it means the current interation will land outside |z|<2, so we
can just terminate the iteration.

Like with the 14-bit arithmetic, numbers are stored as the address of
their entry in the table of squares.  Since this table is in the middle
of the address space, but the numbers we want to be able to calculate
squares of are at the edges of the address space, we invert the
high-order bit of each number for storage.  This is rather simpler than
the complex dance for 14-bit numbers, which saves us enough space in the
kernel to fit in the range checks.

Addition and subtraction on these numbers is fairly easy because only
the high-order bit is affected.  Addition and subtraction both generate
the high-order bit as the EOR of their inputs, so to get the correct
result we just need to invert the high-order bit again after adding or
subtracting an even number of values.

Unfortunately, there's a bug somewhere that is producing rendering
errors.  The Mandelbrot set itself is the right shape, but there are
complicated and asymmetric artifacts in the shading.  There are no
errors on the real axis itself, which suggests the problem is an error
in the handling of zi, somewhere.