Remove a lot of bit-flips from the kernel
If a number won't be squared then we can store it with whatever top bit
is most convenient.  This requires a bit of care to get the right number
of inversions at the end, but I think I've got it right.

This gets us to with 1% of the performance of the 14-bit version.