Correct rounding in the table of squares
BeebAsm's usual rounding mode is to round down, which mean that some
entries in the table of squares had an error of almost 1 ulp and all
the errors were in the same direction.  The new code makes the
conversion to integer explicit and adds a suitable offset first to
give us round-to-nearest instead.