Fix magnitude of of r and i as displayed in the shell
The shell's translation from floating to fixed point wasn't updated in
0ce0461b832f5c23c21dec0e931521a1f62166b2 when the fixed point was moved,
which means that since then the displayed centre co-ordinates have been
wrong by a factor of 2.

I've kept the scroll step at a quarter of the screen even though it was
half a screen before the above change, because I think a quarter screen
is nicer now we don't force a full re-render on scrolling.