Make the "scale" variable in the shell more sensible
The "scale" variable in the shell is now the value that it displays as
"s=", which in turn is the width and height of the displayed image.
This makes the code a bit less confusing, and also makes the cursor
display in the right place since the cursor code already assumed that
was what "scale" meant.