Place limits on how far the user can scroll
There's nothing interesting beyond +/-2, so make that the limit for the
centre of the picture.  This avoids showing the artifacts of exceeding
the fixed-point range.  It also conveniently means that the lack of a
range check on zr and zi at the start of the kernel is now harmless
since it's not possible to scroll to somewhere where the initial zi or
zr is out of range.